requirements.txt
# Core Dependencies
astrapy>=1.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
questionary>=2.0.0
pandas>=2.0.0

# Optional: For embedding generation if needed
sentence-transformers>=2.2.0

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
